zkevm-circuits: add `test-circuits` feature flag
Add an additional feature flag that exports the Circuit implementation for
sub-circuits. This prevents overhead and crossbuild errors when building with the `test` feature.

- evm circuit is an exception because of EXECUTION_STEP_HEIGHT (always exported)
- pi circuit is exported via PiTestCircuit because of associated types
