+++ zkevm-circuits/Cargo.toml
@@ -49,3 +49,4 @@ pretty_assertions = "1.0.0"
 [features]
 default = []
 test = ["ethers-signers", "mock"]
+test-circuits = []
